changes after codereview

+++ events/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render_to_response
 from django.template import RequestContext
 from django import http
+from django.db import transaction
 from django.core.urlresolvers import reverse
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
@@ -11,59 +12,61 @@ from events.models import *
 
 def event_details(request, event_id):
     event = get_object_or_404(Event, id=event_id)
-    all_attractions = Attraction.objects.all().filter(event=event)
-    organizer = Organizer.objects.get(event=event)
+    all_attractions = Attraction.objects.filter(event=event)
     return render_to_response("events/event_details.html",
                               {"event": event,
-                               "all_attractions": all_attractions,
-                               "organizer": organizer},
+                               "all_attractions": all_attractions},
                               context_instance=RequestContext(request))
 
 
 @login_required
+@transaction.commit_on_success
 def add_event(request):
     if request.method == 'POST':
-        form = AddEventForm(request.POST)
-        organizer_form = AddOrganizerForm(request.POST)
+        form = AddEventForm(request.POST, prefix='event')
+        organizer_form = AddOrganizerForm(request.POST, prefix='organizer')
         if form.is_valid():
             event = form.save(commit=False)
             event.created_by = request.user
-            event.save()
-            form.save()
             if organizer_form.is_valid():
                 organizer = organizer_form.save(commit=False)
-                organizer.event = event
+                organizer.related_user = request.user
                 organizer.save()
                 organizer_form.save()
+                event.organizer = organizer
+                event.save()
+                form.save()
                 return http.HttpResponseRedirect(reverse(
                     'events.views.event_details',
                     args=(event.id,)))
     else:
-        form = AddEventForm()
-        organizer_form = AddOrganizerForm()
+        form = AddEventForm(prefix='event')
+        organizer_form = AddOrganizerForm(prefix='organizer')
     return render_to_response("events/event_add.html",
                               {"form": form, "organizer_form": organizer_form},
                               context_instance=RequestContext(request))
 
 
 @login_required
+@transaction.commit_on_success
 def edit_event(request, event_id):
     event = get_object_or_404(Event, pk=event_id)
-    organizer = Organizer.objects.get(event=event)
+    organizer = event.organizer
     if request.method == "POST":
-        form = EditEventForm(request.POST, instance=event)
+        form = EditEventForm(request.POST, instance=event, prefix='event')
         organizer_form = AddOrganizerForm(request.POST,
-                                          instance=organizer)
-        if form.is_valid():
+                                          instance=organizer,
+                                          prefix='organizer')
+        if form.is_valid() and organizer_form.is_valid():
             event = form.save()
-            if organizer_form.is_valid():
-                organizer = organizer_form.save()
-                return http.HttpResponseRedirect(reverse(
-                    'events.views.event_details',
-                    args=(event.id,)))
+            organizer = organizer_form.save()
+            return http.HttpResponseRedirect(reverse(
+                'events.views.event_details',
+                args=(event.id,)))
     else:
-        form = EditEventForm(instance=event)
-        organizer_form = AddOrganizerForm(instance=organizer)
+        form = EditEventForm(instance=event, prefix='event')
+        organizer_form = AddOrganizerForm(instance=organizer,
+                                          prefix='organizer')
 
     return render_to_response("events/event_edit.html",
                               {"form": form, "organizer_form": organizer_form,
@@ -74,8 +77,9 @@ def edit_event(request, event_id):
 @login_required
 def delete_event(request, event_id):
     event = get_object_or_404(Event, pk=event_id)
-    event.delete()
-    messages.success(request, 'Event has been deleted.')
+    if request.method == "POST":
+        event.delete()
+        messages.success(request, 'Event has been deleted.')
     return http.HttpResponseRedirect('/')
 
 
@@ -101,7 +105,7 @@ def add_attraction(request, event_id):
 
 def attraction_details(request, event_id, attraction_id):
     attraction = get_object_or_404(Attraction, id=attraction_id)
-    event = get_object_or_404(Event, pk=event_id)
+    event = attraction.event
     return render_to_response("events/attraction_details.html",
                               {"attraction": attraction, "event": event},
                               context_instance=RequestContext(request))
@@ -110,7 +114,7 @@ def attraction_details(request, event_id, attraction_id):
 @login_required
 def edit_attraction(request, event_id, attraction_id):
     attraction = get_object_or_404(Attraction, pk=attraction_id)
-    event = get_object_or_404(Event, pk=event_id)
+    event = attraction.event
     if request.method == "POST":
         form = AddAttractionForm(request.POST, instance=attraction)
         if form.is_valid():
@@ -129,8 +133,9 @@ def edit_attraction(request, event_id, attraction_id):
 @login_required
 def delete_attraction(request, event_id, attraction_id):
     attraction = get_object_or_404(Attraction, pk=attraction_id)
-    event = get_object_or_404(Event, pk=event_id)
-    attraction.delete()
-    messages.success(request, 'Attraction has been deleted.')
+    event = attraction.event
+    if request.method == "POST":
+        attraction.delete()
+        messages.success(request, 'Attraction has been deleted.')
     return http.HttpResponseRedirect(reverse('events.views.event_details',
                                              args=(event.id,)))
